Add benchmarking for formation accuracy
To benchmark, run bench.py while the formation code is running,
which saves the raw poses of the robots in some files. Run graph.py
to generate a graph of the error, or modify for a metric of your
choice!

NOTE: the current error calculation is taken from the old
strategy and needs updating to give the error for the current code
(probably just the target positions are wrong)